needs minimap check

+++ src_bonus/game/movement_bonus.c
@@ -6,7 +6,7 @@
 /*   By: alvega-g <alvega-g@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 17:18:52 by alvega-g          #+#    #+#             */
-/*   Updated: 2024/06/26 16:01:51 by alvega-g         ###   ########.fr       */
+/*   Updated: 2024/06/26 18:42:52 by alvega-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,15 +43,15 @@ static void	ft_move_minimap(t_data *data, double dx, double dy)
 {
 	static double	fx = 0.0;
 	static double	fy = 0.0;
-	int				ix;
-	int				iy;
+	double			ix;
+	double			iy;
 
 	if ((dx == 0 && dy == 0))
 		return ;
 	fx += dx;
 	fy += dy;
-	ix = (int)fx;
-	iy = (int)fy;
+	ix = fx;
+	iy = fy;
 	if (ix != 0)
 	{
 		data->p_x += ix / (W_HEIGHT / 50 + 0.0);
